telegram_bot/requirements.txt
python-telegram-bot==20.6
# Use a compatible OpenAI package version available on PyPI
openai==0.27.10
